requirements.txt
flask==3.0.0
flask-cors==6.0.1
supabase==2.18.1
python-dotenv==1.1.1
gunicorn==21.2.0
